Refactored seen plugin to separate CLI operation into different file
