+++ flexget/plugins/filter/seen.py
@@ -9,6 +9,7 @@ forget (string)
 """
 
 from __future__ import unicode_literals, division, absolute_import
+
 import logging
 from datetime import datetime, timedelta
 
@@ -16,11 +17,10 @@ from sqlalchemy import Column, Integer, DateTime, Unicode, Boolean, or_, select,
 from sqlalchemy.orm import relation
 from sqlalchemy.schema import ForeignKey
 
-from flexget import db_schema, options, plugin
+from flexget import db_schema, plugin
 from flexget.event import event
-from flexget.logger import console
 from flexget.manager import Session
-from flexget.utils.imdb import is_imdb_url, extract_id
+from flexget.utils.imdb import extract_id
 from flexget.utils.sqlalchemy_utils import table_schema, table_add_column
 
 log = logging.getLogger('seen')
@@ -56,7 +56,6 @@ def upgrade(ver, session):
 
 
 class SeenEntry(Base):
-
     __tablename__ = 'seen_entry'
 
     id = Column(Integer, primary_key=True)
@@ -80,7 +79,6 @@ class SeenEntry(Base):
 
 
 class SeenField(Base):
-
     __tablename__ = 'seen_field'
 
     id = Column(Integer, primary_key=True)
@@ -106,9 +104,7 @@ def forget(value):
     :return: count, field_count where count is number of entries removed and field_count number of fields
     """
     log.debug('forget called with %s' % value)
-    session = Session()
-
-    try:
+    with Session() as session:
         count = 0
         field_count = 0
         for se in session.query(SeenEntry).filter(or_(SeenEntry.title == value, SeenEntry.task == value)).all():
@@ -124,9 +120,6 @@ def forget(value):
             log.debug('forgetting %s' % se)
             session.delete(se)
         return count, field_count
-    finally:
-        session.commit()
-        session.close()
 
 
 class FilterSeen(object):
@@ -241,74 +234,19 @@ def db_cleanup(manager, session):
         log.verbose('Removed %d seen fields older than 1 year.' % result)
 
 
-def do_cli(manager, options):
-    if options.seen_action == 'forget':
-        seen_forget(manager, options)
-    elif options.seen_action == 'add':
-        seen_add(options)
-    elif options.seen_action == 'search':
-        seen_search(options)
-
-
-def seen_forget(manager, options):
-    forget_name = options.forget_value
-    if is_imdb_url(forget_name):
-        imdb_id = extract_id(forget_name)
-        if imdb_id:
-            forget_name = imdb_id
-
-    count, fcount = forget(forget_name)
-    console('Removed %s titles (%s fields)' % (count, fcount))
-    manager.config_changed()
-
-
-def seen_add(options):
-    seen_name = options.add_value
-    if is_imdb_url(seen_name):
-        imdb_id = extract_id(seen_name)
-        if imdb_id:
-            seen_name = imdb_id
-
+def add(seen_name):
     with Session() as session:
         se = SeenEntry(seen_name, 'cli_seen')
         sf = SeenField('cli_seen', seen_name)
         se.fields.append(sf)
         session.add(se)
-    console('Added %s as seen. This will affect all tasks.' % seen_name)
 
 
-def seen_search(options):
-    session = Session()
-    try:
-        search_term = '%' + options.search_term + '%'
-        seen_entries = (session.query(SeenEntry).join(SeenField).
-                        filter(SeenField.value.like(search_term)).order_by(SeenField.added).all())
-
-        for se in seen_entries:
-            console('ID: %s Name: %s Task: %s Added: %s' % (se.id, se.title, se.task, se.added.strftime('%c')))
-            for sf in se.fields:
-                console(' %s: %s' % (sf.field, sf.value))
-            console('')
-
-        if not seen_entries:
-            console('No results')
-    finally:
-        session.close()
+def search(value, session=None):
+        return session.query(SeenEntry).join(SeenField).filter(SeenField.value.like(value)).order_by(
+                SeenField.added).all()
 
 
 @event('plugin.register')
 def register_plugin():
     plugin.register(FilterSeen, 'seen', builtin=True, api_ver=2)
-
-
-@event('options.register')
-def register_parser_arguments():
-    parser = options.register_command('seen', do_cli, help='view or forget entries remembered by the seen plugin')
-    subparsers = parser.add_subparsers(dest='seen_action', metavar='<action>')
-    forget_parser = subparsers.add_parser('forget', help='forget entry or entire task from seen plugin database')
-    forget_parser.add_argument('forget_value', metavar='<value>',
-                               help='title or url of entry to forget, or name of task to forget')
-    add_parser = subparsers.add_parser('add', help='add a title or url to the seen database')
-    add_parser.add_argument('add_value', metavar='<value>', help='the title or url to add')
-    search_parser = subparsers.add_parser('search', help='search text from the seen database')
-    search_parser.add_argument('search_term', metavar='<search term>')
